Corregido tipo de variables

+++ soluciones/02_Variables/S02Ejercicio01.java
@@ -12,14 +12,14 @@
 
 public class S02Ejercicio01 {
   public static void main(String[] args) {
-    float x = 144;
-    float y = 999;
+    int x = 144;
+    int y = 999;
     
     System.out.println("x = " + x);
     System.out.println("y = " + y);
     System.out.println("x + y = " + (x + y));
     System.out.println("x - y = " + (x - y));
-    System.out.println("x / y = " + (x / y));
+    System.out.println("x / y = " + ((float)x / (float)y));
     System.out.println("x * y = " + (x * y));
   }
 }
